Clean a little bitthe Scenario API
